# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Danshin Genealogy",
        "description": "Доступ ко всем данным только через API токен,\n        в противном случае будет возвращаться ошибка 401.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/person": {
            "post": {
                "operationId": "e3fa321f1f16f618321269761b9dce45",
                "parameters": [
                    {
                        "name": "is_unavailable",
                        "in": null,
                        "description": "Доступность лица (поддерживается связь, поступает о нём известна информация).",
                        "required": true
                    },
                    {
                        "name": "is_live",
                        "in": null,
                        "description": "Статус живой или нет.",
                        "required": true
                    },
                    {
                        "name": "gender",
                        "in": null,
                        "description": "Пол (из допустимых значений).",
                        "required": true
                    },
                    {
                        "name": "surname",
                        "in": null,
                        "description": "Фамилия.",
                        "required": false
                    },
                    {
                        "name": "old_surname",
                        "in": null,
                        "description": "Прежние фамилии с порядковыми номерами.",
                        "required": false
                    },
                    {
                        "name": "name",
                        "in": null,
                        "description": "Имя.",
                        "required": false
                    },
                    {
                        "name": "patronymic",
                        "in": null,
                        "description": "Отчество.",
                        "required": false
                    },
                    {
                        "name": "has_patronymic",
                        "in": null,
                        "description": "Имеется ли отчество (по умолчанию отчество предполагается).",
                        "required": false
                    },
                    {
                        "name": "birth_date",
                        "in": null,
                        "description": "Дата рождения в формате гггг-мм-дд, вместо неизвестных цифр допустим символ ?",
                        "required": false
                    },
                    {
                        "name": "birth_place",
                        "in": null,
                        "description": "Место рождения.",
                        "required": false
                    },
                    {
                        "name": "death_date",
                        "in": null,
                        "description": "Дата смерти в формате гггг-мм-дд, вместо неизвестных цифр допустим символ ?",
                        "required": false
                    },
                    {
                        "name": "burial_place",
                        "in": null,
                        "description": "Место захоронения.",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Лицо успешно сохранено."
                    },
                    "422": {
                        "description": "Неверные входные данные."
                    }
                }
            }
        },
        "/api/person/{id}": {
            "get": {
                "operationId": "61b1d22e25fb2fecacff38a32de405fd",
                "parameters": [
                    {
                        "name": "id",
                        "in": null,
                        "description": "Id запрашиваемого лица.",
                        "required": true
                    },
                    {
                        "name": "date",
                        "in": null,
                        "description": "Дата в формате гггг-мм-дд,\n                    от которой будет идти отсчёт возраста лица и других сведений.",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Лицо с переданным id."
                    },
                    "404": {
                        "description": "Лицо с запрошенным id не существует."
                    },
                    "422": {
                        "description": "Неверный формат даты, указанной в параметрах запроса."
                    }
                }
            }
        },
        "/api/person/{id}/edit": {
            "get": {
                "operationId": "4a39e053ac92c3619301a49eb0f7386a",
                "parameters": [
                    {
                        "name": "id",
                        "in": null,
                        "description": "Id запрашиваемого лица.",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Лицо с переданным id для редактирования."
                    },
                    "404": {
                        "description": "Лицо с запрошенным id не существует."
                    }
                }
            }
        },
...
    }
}